imporve folder strcutre, add types
