+++ packages/playground/src/components/profile-settings/ssh-key-create-form.tsx
@@ -3,7 +3,7 @@ import { Button, ButtonGroup, Input, Textarea, Icon } from '@harnessio/canary'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
-import { FormFieldSet } from '../index'
+import { FormFieldSet } from '../../index'
 
 const formSchema = z.object({
   identifier: z.string().min(1, { message: 'Please provide a name' }),
