+++ packages/playground/src/index.ts
@@ -62,11 +62,13 @@ export * from './components/form-inputs/common/InputLabel'
 export * from './components/form-inputs/common/InputWrapper'
 export * from './components/webhook-list'
 export * from './components/pagination'
-export * from './components/profile-settings-keys-list'
-export * from './components/profile-settings-tokens-list'
+export * from './components/profile-settings/profile-settings-keys-list'
+export * from './components/profile-settings/profile-settings-tokens-list'
+export * from './components/profile-settings/types'
+
 export * from './components/filter'
 export * from './components/copy-button'
-export * from './components/delete-alert-dialog'
+export * from './components/profile-settings/delete-alert-dialog'
 
 export * from './pages/signin-page'
 export * from './pages/create-project-page'
