+++ packages/playground/src/pages/sandbox-settings-account-keys-page.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react'
 import { Spacer, Text, Button } from '@harnessio/canary'
 import { FormFieldSet, SandboxLayout } from '..'
-import { ProfileKeysList } from '../components/profile-settings-keys-list'
-import { ProfileTokensList } from '../components/profile-settings-tokens-list'
+import { ProfileKeysList } from '../components/profile-settings/profile-settings-keys-list'
+import { ProfileTokensList } from '../components/profile-settings/profile-settings-tokens-list'
 import { mockKeys } from './mocks/profile-settings/mockKeyList'
 import { mockTokens } from './mocks/profile-settings/mockTokensList'
-import { TokenCreateDialog } from '../components/token-create-dialog'
-import { SshKeyCreateDialog } from '../components/ssh-key-create-dialog'
+import { TokenCreateDialog } from '../components/profile-settings/token-create-dialog'
+import { SshKeyCreateDialog } from '../components/profile-settings/ssh-key-create-dialog'
 import { DeleteTokenAlertDialog } from '..'
 
 function SandboxSettingsAccountKeysPage() {
